moving add_public_project_id to shared updating default value for prompt list s3 init session fix
