+++ api/v1/approve_collection.py
@@ -3,7 +3,7 @@ from ...utils.collections import CollectionPublishing
 from pylon.core.tools import log
 from tools import api_tools, auth, config as c
 
-from ...utils.utils import add_public_project_id
+from ....promptlib_shared.utils.utils import add_public_project_id
 
 
 class PromptLibAPI(api_tools.APIModeHandler):
