+++ api/v1/author.py
@@ -4,9 +4,10 @@ from pylon.core.tools import log
 
 from tools import api_tools, auth, config as c
 
-from ...utils.utils import get_author_data, add_public_project_id
+from ...utils.utils import get_author_data
 from ...utils.constants import PROMPT_LIB_MODE
 from ...utils.author import get_stats
+from ....promptlib_shared.utils.utils import add_public_project_id
 
 
 class PromptLibAPI(api_tools.APIModeHandler):
@@ -26,12 +27,14 @@ class PromptLibAPI(api_tools.APIModeHandler):
         if not author:
             return {'error': f'author with id {author_id} not found'}
         try:
-            author_project_id = self.module.context.rpc_manager.timeout(2).projects_get_personal_project_id(author['id'])
+            author_project_id = self.module.context.rpc_manager.timeout(2).projects_get_personal_project_id(
+                author['id'])
             stats = get_stats(author_project_id, author['id'])
             author.update(stats)
         except Empty:
             ...
         return author, 200
+
 
 class API(api_tools.APIBase):
     url_params = api_tools.with_modes([
