+++ api/v1/public_collection.py
@@ -1,10 +1,9 @@
 from pylon.core.tools import log
 from tools import api_tools, auth, config as c
 
+from ....promptlib_shared.utils.utils import add_public_project_id
 from ...utils.constants import PROMPT_LIB_MODE
 from ...utils.collections import get_collection
-from ...utils.utils import add_public_project_id
-
 
 
 class PromptLibAPI(api_tools.APIModeHandler):
