+++ api/v1/public_prompt.py
@@ -2,10 +2,10 @@ import json
 from pylon.core.tools import log
 
 from ...utils.prompt_utils import get_published_prompt_details
-from ...utils.utils import add_public_project_id
 from ...utils.constants import PROMPT_LIB_MODE
 
 from tools import api_tools, auth, config as c
+from ....promptlib_shared.utils.utils import add_public_project_id
 
 
 class PromptLibAPI(api_tools.APIModeHandler):
